+++ packages/highlight/test/textmate.test.cjs
@@ -40,7 +40,8 @@ test("TextMate grammar recognizes representative Skel constructs", async () => {
     ["    method getUser {", "entity.name.function.method.skel"],
     ["// contract comment", "comment.line.double-slash.skel"],
     ["@desc(\"User contract\")", "entity.name.function.decorator.skel"],
-    ["@sensitive", "entity.name.function.decorator.skel"]
+    ["@sensitive", "entity.name.function.decorator.skel"],
+    ["@deprecated(\"Use Profile instead\")", "entity.name.function.decorator.skel"]
   ];
 
   let ruleStack = textmate.INITIAL;